# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openrouter
    description: OpenRouter AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openrouter/conversations:
    get:
      operationId: listOpenrouterConversations
      tags: [openrouter]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterConversation"
    post:
      operationId: createOpenrouterConversation
      tags: [openrouter]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenrouterConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversation"
  /openrouter/conversations/{id}:
    get:
      operationId: getOpenrouterConversation
      tags: [openrouter]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
    delete:
      operationId: deleteOpenrouterConversation
      tags: [openrouter]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/messages/{messageId}:
    patch:
      operationId: updateOpenrouterMessage
      tags: [openrouter]
      summary: Update the content of an existing message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpenrouterMessageBody"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
    delete:
      operationId: deleteOpenrouterMessage
      tags: [openrouter]
      summary: Delete an existing message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/messages/{messageId}/regenerate:
    post:
      operationId: regenerateOpenrouterMessage
      tags: [openrouter]
      summary: Regenerate a single message in place using AI completion
      description: |
        Replaces the content of an existing message with a freshly AI-generated paragraph.
        The AI sees only the messages that come BEFORE this one as context, then writes a
        new paragraph that fits at this position. The message's role and createdAt are preserved.
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateOpenrouterMessageBody"
      responses:
        "200":
          description: Regenerated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterMessage"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
        "500":
          description: AI completion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/completions:
    post:
      operationId: createOpenrouterCompletion
      tags: [openrouter]
      summary: Generate a single AI completion (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenrouterMessageBody"
      responses:
        "200":
          description: AI-generated answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterCompletion"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
        "429":
          description: Upstream rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
        "500":
          description: Upstream or internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/conversations/{id}/ai-turn:
    post:
      operationId: triggerOpenrouterAiTurn
      tags: [openrouter]
      summary: Stream an AI assistant response based on existing history (SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerOpenrouterAiTurnBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
        "400":
          description: Invalid request or no messages to respond to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/conversations/{id}/messages:
    get:
      operationId: listOpenrouterMessages
      tags: [openrouter]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterMessage"
    post:
      operationId: sendOpenrouterMessage
      tags: [openrouter]
      summary: Send a user message; optionally stream an AI response (SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenrouterMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OpenrouterConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenrouterMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        language:
          type: ["string", "null"]
          description: BCP-47 language tag of the message content (used to pick the right TTS voice).
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenrouterConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenrouterMessageBody:
      type: object
      properties:
        content:
          type: string
        model:
          type: string
        maxTokens:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        apiKey:
          type: ["string", "null"]
        apiUrl:
          type: ["string", "null"]
        skipAiCompletion:
          type: boolean
        language:
          type: ["string", "null"]
          description: BCP-47 language tag of the user message; persisted so TTS playback later uses the correct voice.
      required:
        - content
    OpenrouterConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenrouterMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    UpdateOpenrouterMessageBody:
      type: object
      properties:
        content:
          type: string
        language:
          type: ["string", "null"]
          description: BCP-47 language tag of the updated content (overrides the saved language when provided).
      required:
        - content
    TriggerOpenrouterAiTurnBody:
      type: object
      properties:
        model:
          type: string
        maxTokens:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        apiKey:
          type: ["string", "null"]
        apiUrl:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
          description: BCP-47 language code instructing the AI which language to respond in (e.g. en-US, he-IL).
    RegenerateOpenrouterMessageBody:
      type: object
      properties:
        model:
          type: string
        maxTokens:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        apiKey:
          type: ["string", "null"]
        apiUrl:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
          description: BCP-47 language code instructing the AI which language to respond in.
    OpenrouterCompletion:
      type: object
      properties:
        answer:
          type: string
        model:
          type: string
      required:
        - answer
        - model
    OpenrouterError:
      type: object
      properties:
        error:
          type: string
      required:
        - error